So now you can understand we either use the constructor or we can do the arrow function like so I'm
going to show you one other way.

I'm going to only show you this other way because we've actually kind of technically already used it

inside of this component.

All right I'm going to first reverse the change.

I just did right here.

Back to how it was before.

So very quickly to refresh and just make sure I still see that.

Yes.

OK so I'm back to how I was before.

All right.

Now here's the third way.

Now we've technically already used this on the input right here.

Notice how our input function it has a callback is very much like what we have right here.

But in this case the callback gets passed down into that component using an arrow function.

So rather than defining a method on our class using an arrow function we just pass a arrow function

directly into our prop's right here.

Or Backley directly into the onchange prop so we can do the same thing with on submit right here.

I'm going to wrap on form submit with an arrow function.

So I'll put in my parentheses right there.

I'll put in my arrow and now I'm going to make sure that I invoke on form submit.

So why am I putting on the Princes here.

Remember in this case I'm now defining an arrow function and passing it down into the form.

So when the form gets submitted it's going to call the arrow function.

So it's going to then execute essentially this little segment right here.

Everything from this start on foreign Smidt to the very end.

So it's only going to the form is only going to invoke this arrow function one time.

It's not going to invoke the arrow function and then get back on form submit function and then decide

to invoke that.

Now that's not how it works.

Whenever we use the arrow function here it's only going to call the arrow function one time.

So now we have to make sure that we invoke on form submit.

All right so now we'll try saying this and we'll do the same thing.

Put some text in.

Hit ENTER NOW.

Ok I made a typo here.

Where do I make my typo.

On form submit on forms submit.

I've got a typo.

Oh it's the event object.

My mistake sorry we have to take the event object and make sure that we pass it into on form submit

like so just as we did down here for the onchange.

OK so let's try this now.

I'll put in some text and enter.

And there we go.

All right so that's it.

That's three ways in which we can solve this.

Now in general we're going to go with the route of assigning an arrow function like so that's going

to be the default method that we use to fix this problem.

But in some cases when we decide to inline a callback as we did down here we'll just rely upon this

arrow function to solve the problem for us.