85.3.Solving Context Issues(Section 7: Handling User Input with Forms and Events)
All right so now in this case if we're going to use the arrow function up here we don't have to put

in the Arrow function in the summit handler.

Technically we could but we don't have to do both.

So I'll say this.

On form submit.

And now I'm back to just passing in a reference to the function.

So I do not need to add in the parentheses OK so let's try one more time just to make sure everything

is fixed.

And yep we're looking to.

All right so that's it.

It's probably way more than you ever wanted to know about context and these event handlers.

Like I said you will run into this problem.

So let's take a pause here.

Continue in the next section.