89.2.Axios vs Fetch(Section 8: Making API Requests with React)
So now I'm going to find my app file and it's on my app file.

I'm going to import x CEOs from X sales.

Now traditionally by convention we usually put imports for third party packages or dependencies that

are installed into our project above import statements for files that you and I create.

So by convention I would want to put the import statement for X use above the one for search bar.

There's not a technical reason for us for this.

Again it is just convention.

So now we can make use of X-posts to make request over to the splash API.

We definitely want to make the request.

When a user submits a search term or when this function gets called right here rather than cancel logging

out a term we will replace it with a call to the axioms function before going through that entire process.

Let's take a quick pause because we are going to end up having to put a decent amount of code in here

so quick pause and we'll write this stuff out in the next section.